add set and tree mixins
Adds the set data structure interface, which is really just a layer on
top of the binary search trees.

I also created tree mixins to try to enforce that only trees (and
specifically binary search trees) should be the backing data structure
of sets. That's because you can choose which BST to use for a set.