avl_tree: fix trinode restructuring
I was updating b's parent before rerooting the subtree we're in. Which
is actually updating node's parent, which might lead to node no longer
having a grandparent in the case when we need it. Thus, we wouldn't be
able to reroot the subtree.

I was also dereferencing nullptrs - shoutout bret hart